[codegen][cleanup] Switch safepoint fields to int
Many callers already pass an int, so there is no point in storing as an
uint32_t internally. The style guide also recommends int as the general
data type for integers, even if it's statically known that only positive
values will occur.

R=jkummerow@chromium.org

Bug: v8:12401
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3289157
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#77975}

+++ src/wasm/wasm-code-manager.cc
@@ -437,8 +437,8 @@ void WasmCode::Disassemble(const char* name, std::ostream& os,
   if (safepoint_table_offset_ > 0) {
     SafepointTable table(this);
     os << "Safepoints (size = " << table.size() << ")\n";
-    for (uint32_t i = 0; i < table.length(); i++) {
-      uintptr_t pc_offset = table.GetPcOffset(i);
+    for (int i = 0; i < table.length(); i++) {
+      int pc_offset = table.GetPcOffset(i);
       os << reinterpret_cast<const void*>(instruction_start() + pc_offset);
       os << std::setw(6) << std::hex << pc_offset << "  " << std::dec;
       table.PrintEntry(i, os);
